+++ CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opts.img_text_func` has an obsidian app compatibility, and only accept one path argument.
 - Moved `client:apply_async` -> `api.dir`
 - Deprecate `statusline` options and `vim.g.obsidian` in favor of `footer` options and `vim.b.obsidian_footer`.
+- Moved `client:switch_workspace` -> `Workspace.switch`
 
 ### Fixed